requirements.txt
# Core requirements
python-dotenv>=0.19.0
requests>=2.25.0

# LLM provider clients - install the ones you need
anthropic>=0.5.0  # For Claude API
openai>=0.27.0    # For OpenAI API
langchain>=0.0.200  # For LangChain integration
langchain-openai>=0.0.200  # For LangChain OpenAI integration
langchain-anthropic>=0.0.200  # For LangChain Anthropic integration

# Development requirements
pytest>=7.0.0     # For running tests
black>=22.0.0     # For code formatting
isort>=5.10.0     # For import sorting 
